🐛 Turn on access manager on pi and bug fixing
Fix bug to start monitoring as soon as pi turns on, cause previously when user toggles the start monitoring button on frontend, then only it was starting monitoring, It works well for read write access but if user has only read access, they can't send a message and the monitoring will never start. So changed it to always monitor and remove option from fronend to toggle between starting and stopping monitor.